+++ Client/app/src/main/java/se/ice/client/utility/MockupServer.java
@@ -15,6 +15,10 @@ public class MockupServer implements Domain {
     Map<String, Admin> admins = new HashMap<>(); // All registered admins
     Map<Gcode, Course> courses = new HashMap<>();
 
+    public static final String email = "name@mail.com";
+    public static final String name = "name";
+    public static final String password = "password";
+
     public static void main(String[] arg) {
         System.out.println("hi");
     }
